Lint stuff, add some CI scripts

+++ env.sh
@@ -2,11 +2,18 @@
 
 set -e
 
-function teardown() {
+if [[ "${1}" == "down" ]]; then
     docker compose down --remove-orphans --volumes
-}
+    exit 0
+fi
 
-trap teardown exit
+if [[ "${1}" != "up" ]]; then
+    function teardown() {
+        docker compose down --remove-orphans --volumes
+    }
+
+    trap teardown exit
+fi
 
 docker compose pull
 
@@ -26,4 +33,6 @@ docker compose restart postgres
 
 docker compose up -d
 
-docker compose logs -f -t
+if [[ "${1}" != "up" ]]; then
+    docker compose logs -f -t
+fi
